FreeBSD: Refactor Linux and FreeBSD file operations into new class
- Reduces redundant code between Linux and FreeBSD implementations
- Create abstract base "UnixFileOperations" class from which LinuxFileOperations
  and FreeBSDFileOperations inherit
    - Avoid reimplementing identical methods; override defaults when necessary
