+++ src/linux/Platform.cmake
@@ -21,6 +21,7 @@ set(PLATFORM_SOURCES
     linux/acceleratedcryptographichash_gnutls.cpp
     linux/bootimgcreator_linux.cpp
     linux/secureboot_crypto_linux.cpp
+    unix/file_operations_unix.cpp
     linux/file_operations_linux.cpp
     linux/platformquirks_linux.cpp
 )
